refactor map control
